+++ flutter_inappwebview_android/android/src/main/java/com/pichillilorenzo/flutter_inappwebview_android/chrome_custom_tabs/ChromeCustomTabsChannelDelegate.java
@@ -106,7 +106,7 @@ public class ChromeCustomTabsChannelDelegate extends ChannelDelegateImpl {
         if (chromeCustomTabsActivity != null && chromeCustomTabsActivity.customTabsSession != null) {
           try {
             result.success(chromeCustomTabsActivity.customTabsSession.isEngagementSignalsApiAvailable(new Bundle()));
-          } catch (Exception e) {
+          } catch (Throwable e) {
             result.success(false);
           }
         } else {