+++ flutter_inappwebview_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+- Updated `flutter_inappwebview_platform_interface` version dependency to `1.0.2` 
+- Fixed "Crash when starting ChromeSafariBrowser on Android java.lang.NoSuchMethodError: No virtual method isEngagementSignalsApiAvailable" [#1881](https://github.com/pichillilorenzo/flutter_inappwebview/issues/1881)
+
 ## 1.0.1
 
 - Updated README