+++ flutter_inappwebview_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+
+- Added `PlatformPrintJobController.onComplete` setter
+
 ## 1.0.1
 
 - Updated README