+++ flutter_inappwebview_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+- Updated `flutter_inappwebview_platform_interface` version dependency to `1.0.2`
+
 ## 1.0.0
 
 Initial release.